split multiplex into two files
